refactor: JoinList のコンポーネントを分割
