+++ src/page/photoList/PhotoList.tsx
@@ -2,46 +2,13 @@ import { trpcReact } from '@/trpc';
 import React, { useEffect, useState } from 'react';
 
 import Sidebar from '@/components/SideBar';
-import VrcPhoto from '@/components/ui/VrcPhoto';
 import { Button } from '@/components/ui/button';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { ROUTER_PATHS } from '@/constants';
 import { RefreshCw } from 'lucide-react';
 import { Link } from 'react-router-dom';
+import { PhotoItemList } from './PhotoItemList';
 import { usePhotoItems, useYearMonthList } from './composable';
-
-const WorldInfo = ({
-  vrcWorldId,
-  datetime,
-}: {
-  vrcWorldId: string;
-  datetime: Date;
-}) => {
-  const { data } = trpcReact.getVrcWorldInfoByWorldId.useQuery(vrcWorldId, {
-    staleTime: 1000 * 60 * 5, // キャッシュの有効期限を5分に設定
-    cacheTime: 1000 * 60 * 30, // キャッシュされたデータを30分間メモリに保持
-  });
-  console.log('datetime', data);
-  const date = datetime;
-  const dateToDisplay = `${date.getUTCFullYear()}/${
-    date.getUTCMonth() + 1
-  }/${date.getUTCDate()} ${date.getUTCHours()}:${date.getUTCMinutes()}`;
-  const worldUrl = `https://vrchat.com/home/world/${vrcWorldId}`;
-  const openUrlMutation = trpcReact.openUrlInDefaultBrowser.useMutation();
-  const onClickWorldUrl = () => {
-    openUrlMutation.mutate(worldUrl);
-  };
-  return (
-    <div>
-      <p className="font-medium">
-        <Button onClick={onClickWorldUrl} variant="link" className="px-0">
-          {data?.name ?? vrcWorldId}
-        </Button>
-      </p>
-      <p className="text-sm text-gray-500">Join: {dateToDisplay}</p>
-    </div>
-  );
-};
 
 const PhotoList = () => {
   const { sortedYearMonthList, refetchYearMonthList } = useYearMonthList();
@@ -84,8 +51,6 @@ const PhotoList = () => {
       `${sortedYearMonthList[0].year}-${sortedYearMonthList[0].month}`,
     [sortedYearMonthList],
   );
-
-  const openPhotoPathMutation = trpcReact.openPathOnExplorer.useMutation();
 
   return (
     <div className="h-full grid grid-cols-5 overflow-hidden">
@@ -144,41 +109,7 @@ const PhotoList = () => {
           </div>
         ) : (
           <ScrollArea>
-            <div className="col-span-4">
-              <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 gap-4 mt-5">
-                {photoItemList?.map((item) => {
-                  // item.photoList がある場合は写真一覧を表示する
-                  const photoList = item.photoList.map((photo) => (
-                    <div
-                      key={`photo-container-${photo.datetime.toISOString()}`}
-                      className="col-span-1"
-                    >
-                      <VrcPhoto
-                        key={photo.path}
-                        photoPath={photo.path}
-                        onClickPhoto={() => {
-                          openPhotoPathMutation.mutate(photo.path);
-                        }}
-                      />
-                    </div>
-                  ));
-                  return (
-                    <>
-                      <div
-                        key={`world-info-${item.joinDatetime.toISOString()}`}
-                        className="col-span-full"
-                      >
-                        <WorldInfo
-                          vrcWorldId={item.worldId}
-                          datetime={item.joinDatetime}
-                        />
-                      </div>
-                      {photoList}
-                    </>
-                  );
-                })}
-              </div>
-            </div>
+            <PhotoItemList photoItemList={photoItemList} />
           </ScrollArea>
         )}
       </div>
